Seventh commit
In this commit I drafted and migrated my models for Team, Game, Player,
and Stat. The current models should include all of the necessary data to
make predictions with and also be useful to the user. The rest of the
api module, the machine learning models, and data prep functions still
need to be completed.