Tenth Commit
In this commit I set up an early version of the App.js file so that it
is currently handling the Login and Register pages that I also
completed. In the future I will have to probably change the Login and
Register files to handle OAuth but currently everything should be
working.