Twelfth commit
In this commit I finished my first version of the Home, Player, Search,
and Team pages. Everything should be working as of right now but the
Stats and Games page need to be finished. The App.js file should be
finished but I am holding off just in case I need to make changes.