Thirteenth commit
In this commit I finished my first draft of my front end. I haven't
tested anything yet but everything looks good so far. I modified some of
the files in the previous commits to assure no issues and I finished the
Games and Stats pages. I will begin testing and then move on to the ml
aspect of this app.