Fourteenth commit
In this commit I tested and fixed all bugs througout the app. This
included changed in both the front end and backend. Everything is now
working and ready for the 'betting' aspect to be implemented.
Futhermore, CSS is much needed and also the serializers need to be
modified to include the names in the games and the date in the stats.